# src/backend_go/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Cognix Golang API Documentation",
        "title": "Cognix API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/manage/personas": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return list of allowed personas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Persona"
                ],
                "summary": "return list of allowed personas",
                "operationId": "personas_get_all",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "true for include deleted personas",
                        "name": "archived",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Persona"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "create persona",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Persona"
                ],
                "summary": "create persona",
                "operationId": "persona_create",
                "parameters": [
                    {
                        "description": "persona payload",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parameters.PersonaParam"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Persona"
                        }
                    }
                }
            }
        },
        "/manage/personas/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return persona by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Persona"
                ],
                "summary": "return persona by id",
                "operationId": "persona_get_by_id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "persona id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Persona"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "update persona",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Persona"
                ],
                "summary": "update persona",
                "operationId": "persona_update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "persona id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "persona payload",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parameters.PersonaParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Persona"
                        }
                    }
                }
            }
        },
        "/manage/personas/{id}/{action}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "delete or restore persona",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Persona"
                ],
                "summary": "delete or restore persona",
                "operationId": "persona_delete_restore",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "persona id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "action : delete | restore ",
                        "name": "action",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Persona"
                        }
                    }
                }
            }
        },
        "/tenant/user_info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get user info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenant"
                ],
                "summary": "get user info",
                "operationId": "tenant_get_user_info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                }
            }
        },
        "/tenant/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return list of users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenant"
                ],
                "summary": "return list of users",
                "operationId": "tenant_get_users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.User"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "add new user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenant"
                ],
                "summary": "add new user",
                "operationId": "tenant_add_user",
                "parameters": [
                    {
                        "description": "create user parameter",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parameters.AddUserParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                }
            }
        },
        "/tenant/users/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "edit  user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenant"
                ],
                "summary": "edit user",
                "operationId": "tenant_edit_user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "edit user parameter",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parameters.EditUserParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.ChatMessage": {
            "type": "object",
            "properties": {
                "chat_session_id": {
                    "type": "number"
                },
                "citations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DocumentResponse"
                    }
                },
                "error": {
                    "type": "string"
                },
                "feedback": {
                    "$ref": "#/definitions/model.ChatMessageFeedback"
                },
                "id": {
                    "type": "number"
                },
                "latest_child_message": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_type": {
                    "type": "string"
                },
                "parent_message": {
                    "type": "number"
                },
                "rephrased_query": {
                    "type": "string"
                },
                "time_sent": {
                    "type": "string"
                },
                "token_count": {
                    "type": "integer"
                }
            }
        },
        "model.ChatMessageFeedback": {
            "type": "object",
            "properties": {
                "chat_message_id": {
                    "type": "number"
                },
                "feedback": {
                    "type": "string"
                },
                "id": {
                    "type": "number"
                },
                "up_votes": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.ChatSession": {
            "type": "object",
            "properties": {
                "creation_date": {
                    "type": "string"
                },
                "deleted_date": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "number"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ChatMessage"
                    }
                },
                "one_shot": {
                    "type": "boolean"
                },
                "persona": {
                    "$ref": "#/definitions/model.Persona"
                },
                "persona_id": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "model.DocumentResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "document_id": {
                    "type": "string"
                },
                "id": {
                    "type": "number"
                },
                "link": {
                    "type": "string"
                },
                "message_id": {
                    "type": "number"
                },
                "updated_date": {
                    "type": "string"
                }
            }
        },
...
        "model.JSONMap": {
            "type": "object",
            "additionalProperties": true
        },
        "model.LLM": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "deleted_date": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "endpoint": {
                    "type": "string"
                },
                "id": {
                    "type": "number"
                },
                "last_update": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "model_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "model.Persona": {
            "type": "object",
            "properties": {
                "chat_sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ChatSession"
                    }
                },
                "creation_date": {
                    "type": "string"
                },
                "default_persona": {
                    "type": "boolean"
                },
                "deleted_date": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "description": {
                    "type": "string"
                },
                "display_priority": {
                    "type": "integer"
                },
                "id": {
                    "type": "number"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "last_update": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "llm": {
                    "$ref": "#/definitions/model.LLM"
                },
                "llm_id": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "prompt": {
                    "$ref": "#/definitions/model.Prompt"
                },
                "starter_messages": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
        "model.Prompt": {
            "type": "object",
            "properties": {
                "creation_date": {
                    "type": "string"
                },
                "deleted_date": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "number"
                },
                "last_update": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "name": {
                    "type": "string"
                },
                "persona_id": {
                    "type": "number"
                },
                "system_prompt": {
                    "type": "string"
                },
                "task_prompt": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "model.Tenant": {
            "type": "object",
            "properties": {
                "configuration": {
                    "$ref": "#/definitions/model.JSONMap"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tenant": {
                    "$ref": "#/definitions/model.Tenant"
                },
                "tenant_id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "parameters.AddUserParam": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "parameters.EditUserParam": {
            "type": "object",
            "properties": {
                "role": {
                    "type": "string"
                }
            }
        },
...
        "parameters.PersonaParam": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "model_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "starter_messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/parameters.StarterMessage"
                    }
                },
                "system_prompt": {
                    "type": "string"
                },
                "task_prompt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "parameters.StarterMessage": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "pg.NullTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}